getting some of the CRUD working

+++ webui.py
@@ -2,8 +2,9 @@ import logging
 from fastapi import FastAPI, Depends, Request, Form, status
 from logging.handlers import RotatingFileHandler
 
-from starlette.responses import RedirectResponse
+from starlette.responses import RedirectResponse, JSONResponse
 from starlette.templating import Jinja2Templates
+from starlette.staticfiles import StaticFiles
 
 from sqlalchemy.orm import Session
 import models
@@ -24,6 +25,8 @@ def get_db():
     finally:
         db.close()
 
+app.mount("/static", StaticFiles(directory="static"), name="static")        
+
 # Define a route for the home page
 @app.get('/')
 def home(request: Request, db: Session = Depends(get_db)):
@@ -40,11 +43,19 @@ def first_run(request: Request, db: Session = Depends(get_db)):
 @app.post('/add_movie/')
 def add_movie(request: Request, db: Session = Depends(get_db), video_path: str = Form(...)):
     movie = models.Movie(video_path=video_path)
-    db.add(movie)
-    db.commit()
+    existingMovie = db.query(models.Movie).filter(models.Movie.video_path == video_path).first()
 
-    url = app.url_path_for('home')
-    return RedirectResponse(url=url, status_code=status.HTTP_303_SEE_OTHER)
+    # if existingMovie exists then redirect to update_movie
+
+    if existingMovie:
+        url = app.url_path_for('update_movie', movie_id=existingMovie.id)
+        return RedirectResponse(url=url, status_code=status.HTTP_303_SEE_OTHER)
+    else:        
+        db.add(movie)
+        db.commit()
+
+        url = app.url_path_for('home')
+        return RedirectResponse(url=url, status_code=status.HTTP_303_SEE_OTHER)
 
 @app.post('/update_movie/{movie_id}')
 def update_movie(request: Request, movie_id: int, db: Session = Depends(get_db)):
@@ -58,8 +69,10 @@ def update_movie(request: Request, movie_id: int, db: Session = Depends(get_db))
 @app.post('/delete_movie/{movie_id}')
 def delete_movie(request: Request, movie_id: int, db: Session = Depends(get_db)):
     movie = db.query(models.Movie).filter(models.Movie.id == movie_id).first()
-    db.delete(movie)
-    db.commit()
+    if movie:
+        db.delete(movie)
+        db.commit()
 
-    url = app.url_path_for('home')
-    return RedirectResponse(url=url, status_code=status.HTTP_303_FOUND)
+        return JSONResponse(status_code=status.HTTP_302_FOUND, content={"message": "Movie item deleted successfully"})
+    else:        
+        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"error": "invalid id", "message": "Movie not found"})
